Create seeds for users, restaurants, and reservations; added bcrypt

+++ db/migrate/20170207220653_create_users.rb
@@ -2,7 +2,7 @@ class CreateUsers < ActiveRecord::Migration[5.0]
   def change
     create_table :users do |t|
       t.string :name
-      t.integer :phone_number
+      t.bigint :phone_number
       t.string :email
       t.string :password_digest
